Fix issue where client does network calls on EDT and to the wrong region
